dt intent

+++ SwampJump/Game/Source/Scene.h
@@ -47,11 +47,11 @@ public:
 
 	struct player_struct
 	{
-		int x;
-		int y;
-		int vx = 0;
-		int vy = 0;
-		int ay = 1;
+		float x;
+		float y;
+		float vx = 0;
+		float vy = 0;
+		float ay = 1;
 
 		float w;
 		float h;
